backend/lambda_functions/image_processor/requirements.txt
Pillow
boto3
